Move EVMDomainExecutorDispatch to domain module and publicize domain module
No logical changes.
